Refactor app and move manual relationship tests
